requirements.txt
keras==2.8.0
numpy==1.21.6
Pillow==9.1.1
requests==2.23.0
streamlit==1.9.0
tensorflow==2.8.0
